+++ rig/rig-core/src/providers/openai/responses_api/mod.rs
@@ -968,9 +968,7 @@ where
     T: HttpClientExt + Clone + Default + std::fmt::Debug + 'static,
 {
     /// Use the Completions API instead of Responses.
-    pub fn completions_api(
-        self,
-    ) -> crate::providers::openai::completion::CompletionModel<T> {
+    pub fn completions_api(self) -> crate::providers::openai::completion::CompletionModel<T> {
         super::completion::CompletionModel::with_model(self.client.completions_api(), &self.model)
     }
 }
